+++ example/lib/example5_themes_playground/widgets/panels/bottom_sheet_settings/bottom_sheet_settings.dart
@@ -19,7 +19,7 @@ class BottomSheetSettings extends StatelessWidget {
 
     final String sheetDefaultColorLabel =
         controller.useSubThemes && controller.useFlexColorScheme
-            ? 'default (surface)'
+            ? 'default (surfaceContainerLow)'
             : useMaterial3
                 ? 'default (surface)'
                 : 'default (theme.canvasColor)';